Reworked mocks
