+++ unit/api/healthcheck/test_healthcheck.py
@@ -2,23 +2,135 @@
 Tests the healthcheck API correctly calls the underlying Elasticsearch client
 """
 from unittest import mock
+from unittest.mock import MagicMock
 
 from unit.utils.search_test_app import SearchTestApp
-from unit.elasticsearch.elasticsearch_test_utils import (
-    mock_health_check_client_green,
-    mock_health_check_client_yellow,
-    mock_health_check_client_red,
-    mock_health_check_client_exception,
-    mock_indices_exists_client,
-    mock_indices_not_exists_client
-)
 
-from unit.fasttext.fasttext_test_utils import mock_fasttext_client, mock_unhealthy_fasttext_client
+from unit.elasticsearch.elasticsearch_test_utils import mock_health_response, MockElasticsearchClient
+
+from dp_fasttext.client.testing.mock_client import MockClient, mock_fasttext_client
 
 from dp_conceptual_search.config.config import SEARCH_CONFIG
 from dp_conceptual_search.api.healthcheck.routes import Services, HeathCheckResponse
 from dp_conceptual_search.ons.search.conceptual.client.fasttext_client import FastTextClientService
 from dp_conceptual_search.app.elasticsearch.elasticsearch_client_service import ElasticsearchClientService
+
+
+def mock_indices_exists_client(*args):
+    """
+    Mocks a True result for the indices.exists API
+    :param args:
+    :return:
+    """
+    health_response = mock_health_response("green")
+
+    mock_client = MockElasticsearchClient()
+
+    mock_client.cluster.health = MagicMock(return_value=health_response)
+    mock_client.indices.exists = MagicMock(return_value=True)
+
+    return mock_client
+
+
+def mock_indices_not_exists_client(*args):
+    """
+    Mocks a True result for the indices.exists API
+    :param args:
+    :return:
+    """
+    health_response = mock_health_response("green")
+
+    mock_client = MockElasticsearchClient()
+
+    mock_client.cluster.health = MagicMock(return_value=health_response)
+    mock_client.indices.exists = MagicMock(return_value=False)
+
+    return mock_client
+
+
+def mock_health_check_client_green(*args):
+    """
+    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'green'
+    :return:
+    """
+    # Mock search response
+    health_response = mock_health_response("green")
+
+    # Mock the search client
+    mock_client = MockElasticsearchClient()
+    mock_client.cluster.health = MagicMock(return_value=health_response)
+    mock_client.indices.exists = MagicMock(return_value=True)
+
+    return mock_client
+
+
+def mock_health_check_client_yellow(*args):
+    """
+    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'yellow'
+    :return:
+    """
+    # Mock search response
+    health_response = mock_health_response("yellow")
+
+    # Mock the search client
+    mock_client = MockElasticsearchClient()
+    mock_client.cluster.health = MagicMock(return_value=health_response)
+    mock_client.indices.exists = MagicMock(return_value=True)
+
+    return mock_client
+
+
+def mock_health_check_client_red(*args):
+    """
+    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'red'
+    :return:
+    """
+    # Mock search response
+    health_response = mock_health_response("red")
+
+    # Mock the search client
+    mock_client = MockElasticsearchClient()
+    mock_client.cluster.health = MagicMock(return_value=health_response)
+    mock_client.indices.exists = MagicMock(return_value=True)
+
+    return mock_client
+
+
+def mock_health_check_client_exception(*args):
+    """
+    Returns a mock Elasticsearch client for healthchecks which raises an exception on call
+    :return:
+    """
+    # Mock the search client
+    mock_client = MockElasticsearchClient()
+    mock_client.cluster.health = MagicMock()
+    mock_client.cluster.health.side_effect = Exception("Mock exception")
+    mock_client.indices.exists = MagicMock(return_value=True)
+
+    return mock_client
+
+
+def raise_exception():
+    """
+    Simple method to raise an exception
+    :return:
+    """
+    raise Exception("Mock exception")
+
+
+def mock_unhealthy_fasttext_client():
+    """
+    Mocks an unhealthy dp-fasttext client
+    :return:
+    """
+    # Initialise the MockClient and mock the 'get' and 'post' methods
+    client = MockClient()
+    client.get = MagicMock()
+
+    # Set side effect to new method so we can preserve calling arguments
+    client.get.side_effect = raise_exception
+
+    return client
 
 
 class HealthCheckTestCase(SearchTestApp):
