+++ unit/elasticsearch/elasticsearch_test_utils.py
@@ -7,109 +7,9 @@ def mock_search_client(*args):
     Returns a mock Elasticsearch client for search
     :return:
     """
-    # Mock search response
-    search_response = mock_search_response()
-    health_response = mock_health_response("green")
-
     # Mock the search client
     mock_client = MockElasticsearchClient()
-    mock_client.search = MagicMock(return_value=search_response)
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=True)
-
-    return mock_client
-
-
-def mock_indices_exists_client(*args):
-    """
-    Mocks a True result for the indices.exists API
-    :param args:
-    :return:
-    """
-    health_response = mock_health_response("green")
-
-    mock_client = MockElasticsearchClient()
-
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=True)
-
-    return mock_client
-
-
-def mock_indices_not_exists_client(*args):
-    """
-    Mocks a True result for the indices.exists API
-    :param args:
-    :return:
-    """
-    health_response = mock_health_response("green")
-
-    mock_client = MockElasticsearchClient()
-
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=False)
-
-    return mock_client
-
-
-def mock_health_check_client_green(*args):
-    """
-    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'green'
-    :return:
-    """
-    # Mock search response
-    health_response = mock_health_response("green")
-
-    # Mock the search client
-    mock_client = MockElasticsearchClient()
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=True)
-
-    return mock_client
-
-
-def mock_health_check_client_yellow(*args):
-    """
-    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'yellow'
-    :return:
-    """
-    # Mock search response
-    health_response = mock_health_response("yellow")
-
-    # Mock the search client
-    mock_client = MockElasticsearchClient()
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=True)
-
-    return mock_client
-
-
-def mock_health_check_client_red(*args):
-    """
-    Returns a mock Elasticsearch client for healthchecks with a cluster status of 'red'
-    :return:
-    """
-    # Mock search response
-    health_response = mock_health_response("red")
-
-    # Mock the search client
-    mock_client = MockElasticsearchClient()
-    mock_client.cluster.health = MagicMock(return_value=health_response)
-    mock_client.indices.exists = MagicMock(return_value=True)
-
-    return mock_client
-
-
-def mock_health_check_client_exception(*args):
-    """
-    Returns a mock Elasticsearch client for healthchecks which raises an exception on call
-    :return:
-    """
-    # Mock the search client
-    mock_client = MockElasticsearchClient()
-    mock_client.cluster.health = MagicMock()
-    mock_client.cluster.health.side_effect = Exception("Mock exception")
-    mock_client.indices.exists = MagicMock(return_value=True)
+    mock_client.search = MagicMock(return_value=mock_search_response())
 
     return mock_client
